add library to replace base Extensions and use it's in all project's
