+++ Algoritms/HomeWork8/Helper/Helper.csproj
@@ -1,12 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <OutputType>Exe</OutputType>
         <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Helper\Helper.csproj" />
-    </ItemGroup>
 
 </Project>
