+++ Algoritms/HomeWork8/HomeWork8.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SortHomeWork", "SortHomeWor
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LessonsTests", "LessonsTests\LessonsTests.csproj", "{B5A9E7B3-9922-467A-B67D-48338603CC07}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Helper", "Helper\Helper.csproj", "{885D0D1E-F3BF-4B8A-8296-B3A9BF4C7559}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{B5A9E7B3-9922-467A-B67D-48338603CC07}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B5A9E7B3-9922-467A-B67D-48338603CC07}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B5A9E7B3-9922-467A-B67D-48338603CC07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{885D0D1E-F3BF-4B8A-8296-B3A9BF4C7559}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{885D0D1E-F3BF-4B8A-8296-B3A9BF4C7559}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{885D0D1E-F3BF-4B8A-8296-B3A9BF4C7559}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{885D0D1E-F3BF-4B8A-8296-B3A9BF4C7559}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
