+++ Algoritms/HomeWork8/SortHomeWork/SortHomeWork.csproj
@@ -5,4 +5,8 @@
         <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\Helper\Helper.csproj" />
+    </ItemGroup>
+
 </Project>
